Исправлена обработка заказа

+++ payment-service/domain/models/status-response.go
@@ -1,7 +1,7 @@
 package models
 
 type OrderStatusResponse struct {
-	OrderID int    `json:"order_id"`
+	OrderID int64  `json:"order_id"`
 	Order   Order  `json:"order"`
 	Message string `json:"message"`
 }
